Add tango-runtime for basic log read/write functionality
The tango-runtime function creates an atom that has access to a log and a
Tango Object registry.

update-helper is used to allow writes to the log.

query-helper is used to read the log and builds the registered Tango Objects
in memory.

The test file that was created is not really in the unit test format, but was
useful for debugging in the repl, so I included it.