Add SharedLog protocol and modify the Tango Runtime to use it
I wrapped the functionality previously provided in tango.log.atom with a
clojure record that implements the SharedLog protocol, and also provided a
constructor to make it easy to get an initialized Log record for the runtime.

tango-runtime was updated to use the SharedLog interface to do it's work, and
given access to tango.log.atom.log-record to provide clients with a function
to create an in-memory log from the runtime namespace.

tango.runtime-test was updated to use these changes to know less about the
specific log.