Checkpoint checkin for lots of work for transaction handling
The changes were getting pretty big, so I wanted to just get a snapshot.
The only working object is the tango-register, and the runtime and
transaction-runtime are both in a fluid state (but working...)

I added a simulator to the test code so that it would be easier
to share tests with other impelemenations for more correctness
once it is fully done.

Again, this is incomplete and will be cleaned up.