Reorganize runtime namespaces
This commit is just to clean up some stuff.

The TransactionTangoRuntime record was moved over to transaction-runtime.clj
now that the dust has settled a little in this code.

I also move validate-commit logic to runtime.core so that it is available
to both runtime namespaces.

The simulator was cleaned up a little to make it easier to run test
cases, using the run-commands function.