Track aborted transaction commits in the runtime
Just added a list of aborted commits so that it is available to
the runtime. This makes it so that there is a way for a user to
tell that an abort decision was made, and this is useful in the
simulator to examine the state of the runtime in the case of it
behaving unexpectedly.

The simulator run-commands function was updated to return true if
all of the traces are observed in the runtime, and it will now
return the state of the main runtime if there are any failures,
for analysis.