Add traces to provide tests to run with the simulator
NOTE: Not all of the tests pass yet, there are some consequences to taking
a snapshot that I am still working out. The transaction runtime code was
updated to take a snapshot and try to apply changes to the clone in order
to wait as late as possible for first read to be up to date.

Still need to work out bugs, but this is a good checkpoint checkin.